==> rmds/07_visualise_expanded_model.Rmd <==
---
title: "Visualise expanded model"
author: "David Henderson, Tim Churches, Sumeet Kalia"
date: "6-11 September 2018"
output:
  html_document:
    df_print: paged
    theme: spacelab
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, cache=TRUE, fig.width=8, fig.height=6)
options(digits = 3)
```

# Intro


##Packages

```{r libs}
library(tidyverse)
library(magrittr)
library(feather)
library(effects)
theme_set(theme_minimal(base_size = 16))
```

##Data

```{r load_data, echo=FALSE}
expglm_df <- read_feather(here::here("assets/clean_data/expglm_df.feather"))
expglm_df
#intglm_df <- read_feather(here::here("assets/clean_data/intglm_df.feather"))
#intglm_df
# intglm <- load(here::here("assets/clean_data/intglm.rds"))
```


```{r plot-func, echo=FALSE}
orplot <- function(df, slice_range, ref_level="", prefix="", ylab="", sorted=FALSE) {
    df %<>% 
      slice(slice_range) #Start with age groups only
    if (stringr::str_length(prefix) > 0) {
      y_labels <- stringr::str_remove(df$term, prefix) 
    } else {
      y_labels <- NULL
    }
    if (stringr::str_length(ylab) > 0) {
      title <- paste("Mortality odds ratio -", ylab)
    } else {
      title <- "Mortality odds ratio"
    }      
    if(sorted) {
        g <- df %>%  ggplot(aes(or, fct_reorder(term, or)))
    } else {
        g <- df %>%  ggplot(aes(or, term))
    }
      g <- g + geom_vline(aes(xintercept = 1), 
                 size = 0.25, 
                 linetype = "dashed") +
      geom_errorbarh(aes(xmax = ci95_upper, xmin = ci95_lower), 
                    size = 0.5,
                    height = 0.2,
                    colour = "black") +
      geom_point(size = 2, colour = "#4477AA") +
      scale_x_log10(breaks = scales::pretty_breaks())
      if (!is.null(y_labels)) {
        g <- g + scale_y_discrete(labels = y_labels)
      }
      g + labs(title = title,
          subtitle = paste("Reference:",  ref_level),
           y = ylab,
           x = "Odds Ratio (log scale)")
}
```

## Expanded model

```{r expglm_age_grp}
orplot(expglm_df, slice_range=2:15, ref_level="19-24", prefix="age_grp", ylab="Age groups")
```

```{r expglm_sex}
orplot(expglm_df, slice_range=16, ref_level="Female", prefix="sex", ylab="Sex")
```

```{r expglm_loinc_decile}
orplot(expglm_df, slice_range=33:41, ref_level="Decile 10 - highest", prefix="loinc_decile", ylab="Income deciles")
```


```{r expglm_province}
orplot(expglm_df, slice_range=43:54, ref_level="Ontario", sorted=TRUE, prefix="province", ylab="Province")
```

```{r expglm_rur_urb}
orplot(expglm_df, slice_range=42, ref_level="Urban", prefix="rur_urb", ylab="Rural/urban")
```


```{r expglm_gen}
orplot(expglm_df, slice_range=19:20, ref_level="Born in Canada, both parents born in Canada", prefix="generation", ylab="Migrant generation")
```





```{r expglm_ed}
orplot(expglm_df, slice_range=21:32, ref_level="High school", sorted=TRUE, prefix="education", ylab="Education")
```




```{r expglm_mar}
orplot(expglm_df, slice_range=55:58, ref_level="Married", sorted=TRUE, prefix="mar_stat", ylab="Marital status")
```



```{r expglm_kids, eval=FALSE}
orplot(expglm_df, slice_range=83:84, ref_level="No kids", prefix="no_kids", ylab="No. of children")
```

```{r expglm_ab_id_dichot, eval=FALSE}
orplot(expglm_df, slice_range=17, ref_level="No aboriginal ID", prefix="ab_id_dichot", ylab="Aboriginal ID")
```

```{r expglm_visible_minority, eval=FALSE}
orplot(expglm_df, slice_range=27:37, ref_level="Not a visible minority (except indigenous)", prefix="vis_minority", ylab="Visible minority (except indigenous)", sorted=TRUE)
```



```{r expglm_employ}
orplot(expglm_df, slice_range=17:18, ref_level="Full-time work", prefix="employ_status", ylab="Employment status")
```

```{r expglm_first_language, eval=FALSE}
orplot(expglm_df, slice_range=38:40, ref_level="English", prefix="first_lang", ylab="First language")
```

```{r expglm_dl_difficulty, eval=FALSE}
orplot(expglm_df, slice_range=18:20, ref_level="No ADL difficulties", prefix="adl_difficulty", ylab="Difficulty in activities of daily living")
```




## Model with interactions

```{r interactionglm_age_grp, options, eval=FALSE}
#orplot(intglm_df, slice_range=2:15, ref_level="19-24", prefix="age_grp", ylab="Age groups")
```

```{r interactionglm_sex, options, eval=FALSE}
orplot(intglm_df, slice_range=16, ref_level="Female", prefix="sex", ylab="Sex")
```

```{r interactionglm_ab_id_dichot, options, eval=FALSE}
orplot(intglm_df, slice_range=17, ref_level="No aboriginal ID", prefix="ab_id_dichot", ylab="Aboriginal ID")
```

```{r interactionglm_loinc_decile, options, eval=FALSE}
orplot(intglm_df, slice_range=57:65, ref_level="Decile 10 - highest", prefix="loinc_decile", ylab="Income deciles")
```

```{r interactionglm_province, options, eval=FALSE}
orplot(intglm_df, slice_range=67:78, ref_level="Ontario", sorted=TRUE, prefix="province", ylab="Province")
```

```{r interactionglm_gen, options, eval=FALSE}
orplot(intglm_df, slice_range=43:44, ref_level="Born in Canada, both parents born in Canada", prefix="generation", ylab="Migrant generation")
```

```{r interactionglm_ed, options, eval=FALSE}
orplot(intglm_df, slice_range=45:56, ref_level="No formal education", sorted=TRUE, prefix="education", ylab="Education")
```


```{r interactionglm_mar, options, eval=FALSE}
orplot(intglm_df, slice_range=79:80, ref_level="Married", sorted=TRUE, prefix="mar_stat", ylab="Marital status")
```

```{r interactionglm_kids, options, eval=FALSE}
orplot(intglm_df, slice_range=83:84, ref_level="No kids", prefix="no_kids", ylab="No. of children")
```

```{r interactionglm_visible_minority, options, eval=FALSE}
orplot(intglm_df, slice_range=27:37, ref_level="Not a visible minority (except indigenous)", prefix="vis_minority", ylab="Visible minority (except indigenous)", sorted=TRUE)
```

```{r interactionglm_employ, options, eval=FALSE}
orplot(intglm_df, slice_range=41:42, ref_level="Full-time work", prefix="employ_status", ylab="Employment status")
```

```{r interactionglm_first_language, options, eval=FALSE}
orplot(intglm_df, slice_range=38:40, ref_level="English", prefix="first_lang", ylab="First language")
```

```{r interactionglm_dl_difficulty, options, eval=FALSE}
orplot(intglm_df, slice_range=18:20, ref_level="None", prefix="adl_difficulty", ylab="Difficulty in activities of daily living")
```

## Effect plots
