==> rmds/08_dashboard.Rmd <==
---
title: "Logisitic Regression of Mortality in Canada"
author: "Sumeet Kalia, Tim Churches, David Henderson"
output: 
  flexdashboard::flex_dashboard:
    theme: readable
---
    
Column 
-------------------------------------
    
###
    
```{r map, echo=FALSE, out.width = '100%'}
knitr::include_graphics(here::here("assets/Sumeet/mort-rate.PNG"))
```


Column {.tabset}
-------------------------------------

###The process

* We attempted to create a reproducible work flow using R, RMarkdown, and GitHub
* SAS output incorporated with .png (SAS *can* be run from RMarkdown but we didn't do this)
* A website with all (currently untidy) code is hosted on GitHub pages here https://bit.ly/2Nc11GR
* The full GitHub repo is here https://github.com/CBDRH/ipdln_hackathon_2018


### The Model

We fitted the following logistic regression model to generate the adjusted mortality rate:

**R Code**

```{r r_code, echo=TRUE, eval=FALSE}
expglm <- speedglm(dead ~ 
                     age_grp + sex + loinc_decile + rur_urb +
                     province + education + employ_status +
                     ab_id_dichot + mar_stat + generation,
                   family = binomial(), data = clean_data)`
```

**SAS code**

```{sas, echo=TRUE, eval = FALSE}
model dead(ref='Not Dead')= age_grp   sex loinc_decile rur_urb province  education	employ_status ab_id_dichot mar_stat  generation   ;
```


### Province
    
```{r}
knitr::include_graphics(here::here("rmds/07_visualise_expanded_model_files/figure-html/expglm_province-1.png"))
```
                     
### Age groups

```{r}
knitr::include_graphics(here::here("rmds/07_visualise_expanded_model_files/figure-html/expglm_age_grp-1.png"))
```   

### Sex

```{r}
knitr::include_graphics(here::here("rmds/07_visualise_expanded_model_files/figure-html/expglm_sex-1.png"))
```   


### Rural/Urban
    
```{r}
knitr::include_graphics(here::here("rmds/07_visualise_expanded_model_files/figure-html/expglm_rur_urb-1.png"))
```
 
### Education
    
```{r}
knitr::include_graphics(here::here("rmds/07_visualise_expanded_model_files/figure-html/expglm_ed-1.png"))
```

### Employment status

```{r}
knitr::include_graphics(here::here("rmds/07_visualise_expanded_model_files/figure-html/expglm_employ-1.png"))
```

###Migrant generation

```{r}
knitr::include_graphics(here::here("rmds/07_visualise_expanded_model_files/figure-html/expglm_gen-1.png"))
```

###Income decile

```{r}
knitr::include_graphics(here::here("rmds/07_visualise_expanded_model_files/figure-html/expglm_loinc_decile-1.png"))
```

###Marital Status

```{r}
knitr::include_graphics(here::here("rmds/07_visualise_expanded_model_files/figure-html/expglm_mar-1.png"))
```


### Suggestions

If using `synthpop` to create synthetic data:

* document the generation process (show us the code!)
    * the order of column generation matters!
    * interactions are only synthesised for columns to the left of each column
* provide replicates --- multiple versions of the synthetic dataset
    * stochastic error in the mean of the synthetic estimates can thus be reduced
